ev_auth_ms/requirements.txt
Django==1.11.28
djangorestframework==3.9.1
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
httpie==1.0.3
gunicorn==19.6.0
